+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.0.2] - 2021-06-30.
+
+* add line_view.
+* add line_or_line_view.
+* add appbar_search_view.
+* add form_input_view.
+* add title_require_view.
+* add form_presenter to check validate.
+* add loadmodel_presenter to check validate.
+
+
 ## [0.0.1] - 2021-06-13.
 
 * Initial version, created by dudv.